+++ stacker/commands/stacker/__init__.py
@@ -34,12 +34,10 @@ class Stacker(BaseCommand):
             options.provider = default.Provider(region=options.region)
         options.context = Context(
             environment=options.environment,
-            parameters=copy.deepcopy(options.parameters),
-            # We use
-            # set_default(get_context_kwargs=subcommand.get_context_kwargs) so
-            # the subcommand can provide any specific kwargs to the Context
-            # that it wants. We need to pass down the options so it can
-            # reference any arguments it defined.
+            variables=copy.deepcopy(options.variables),
+            logger_type=self.logger_type,
+            # Allow subcommands to provide any specific kwargs to the Context
+            # that it wants.
             **options.get_context_kwargs(options)
         )
         options.context.load_config(options.config.read())