+++ stacker/tests/blueprints/test_base.py
@@ -4,23 +4,25 @@ from mock import MagicMock
 from troposphere import (
     Base64,
     Ref,
+    s3,
 )
 
 from stacker.blueprints.base import (
     Blueprint,
     CFNParameter,
     build_parameter,
-    get_local_parameters,
+    validate_allowed_values,
     validate_variable_type,
-    resolve_variable
+    resolve_variable,
 )
 from stacker.blueprints.variables.types import (
+    CFNNumber,
     CFNString,
     EC2AvailabilityZoneNameList,
+    TroposphereType,
 )
 from stacker.exceptions import (
     InvalidLookupCombination,
-    MissingLocalParameterException,
     MissingVariable,
     UnresolvedVariable,
     UnresolvedVariables,
@@ -39,31 +41,6 @@ def mock_lookup_handler(value, provider=None, context=None, fqn=False,
 
 
 register_lookup_handler("mock", mock_lookup_handler)
-
-
-class TestLocalParameters(unittest.TestCase):
-    def test_default_parameter(self):
-        parameter_def = {"Param1": {"default": 0}}
-        parameters = {}
-
-        local = get_local_parameters(parameter_def, parameters)
-        self.assertEquals(local["Param1"], 0)
-
-    def test_missing_required(self):
-        parameter_def = {"Param1": {"default": 0}, "Param2": {}}
-        parameters = {}
-
-        with self.assertRaises(MissingLocalParameterException) as cm:
-            get_local_parameters(parameter_def, parameters)
-
-        self.assertEquals("Param2", cm.exception.parameter)
-
-    def test_supplied_parameter(self):
-        parameter_def = {"Param1": {"default": 0}, "Param2": {}}
-        parameters = {"Param1": 1, "Param2": 2}
-
-        local = get_local_parameters(parameter_def, parameters)
-        self.assertEquals(parameters, local)
 
 
 class TestBuildParameter(unittest.TestCase):
@@ -187,6 +164,35 @@ class TestVariables(unittest.TestCase):
             resolve_variable(var_name, var_def, provided_variable,
                              blueprint_name)
 
+    def test_resolve_variable_troposphere_list_type(self):
+        var_name = "testVar"
+        var_def = {"type": TroposphereType(s3.Bucket, many=True)}
+        bucket_defs = {
+            "FirstBucket": {"BucketName": "some-bucket"},
+            "SecondBucket": {"BucketName": "some-other-bucket"},
+        }
+        provided_variable = Variable(var_name, bucket_defs)
+        blueprint_name = "testBlueprint"
+
+        value = resolve_variable(var_name, var_def, provided_variable,
+                                 blueprint_name)
+        for bucket in value:
+            self.assertTrue(isinstance(bucket, s3.Bucket))
+            self.assertEqual(bucket.properties, bucket_defs[bucket.title])
+
+    def test_resolve_variable_troposphere_type(self):
+        var_name = "testVar"
+        var_def = {"type": TroposphereType(s3.Bucket)}
+        provided_variable = Variable(var_name, {"MyBucket": {"BucketName":
+                                                             "some-bucket"}})
+        blueprint_name = "testBlueprint"
+
+        value = resolve_variable(var_name, var_def, provided_variable,
+                                 blueprint_name)
+        self.assertTrue(isinstance(value, s3.Bucket))
+        self.assertEqual(value.properties["BucketName"], "some-bucket")
+        self.assertEqual(value.title, "MyBucket")
+
     def test_resolve_variable_provided_resolved(self):
         var_name = "testVar"
         var_def = {"type": str}
@@ -197,6 +203,20 @@ class TestVariables(unittest.TestCase):
         value = resolve_variable(var_name, var_def, provided_variable,
                                  blueprint_name)
         self.assertEqual(value, "1")
+
+    def test_resolve_variable_allowed_values(self):
+        var_name = "testVar"
+        var_def = {"type": str, "allowed_values": ["allowed"]}
+        provided_variable = Variable(var_name, "not_allowed")
+        blueprint_name = "testBlueprint"
+        with self.assertRaises(ValueError):
+            resolve_variable(var_name, var_def, provided_variable,
+                             blueprint_name)
+
+        provided_variable = Variable(var_name, "allowed")
+        value = resolve_variable(var_name, var_def, provided_variable,
+                                 blueprint_name)
+        self.assertEqual(value, "allowed")
 
     def test_resolve_variable_validator_valid_value(self):
         def triple_validator(value):
@@ -392,6 +412,19 @@ class TestVariables(unittest.TestCase):
         variables = blueprint.get_variables()
         self.assertTrue(isinstance(variables["Param1"], CFNParameter))
 
+    def test_resolve_variables_cfn_number(self):
+        class TestBlueprint(Blueprint):
+            VARIABLES = {
+                "Param1": {"type": CFNNumber},
+            }
+
+        blueprint = TestBlueprint(name="test", context=MagicMock())
+        variables = [Variable("Param1", 1)]
+        blueprint.resolve_variables(variables)
+        variables = blueprint.get_variables()
+        self.assertTrue(isinstance(variables["Param1"], CFNParameter))
+        self.assertEqual(variables["Param1"].value, "1")
+
     def test_resolve_variables_cfn_type_list(self):
         class TestBlueprint(Blueprint):
             VARIABLES = {
@@ -405,7 +438,7 @@ class TestVariables(unittest.TestCase):
         self.assertTrue(isinstance(variables["Param1"], CFNParameter))
         self.assertEqual(variables["Param1"].value, ["us-east-1", "us-west-2"])
         self.assertEqual(variables["Param1"].ref.data, Ref("Param1").data)
-        parameters = blueprint.get_cfn_parameters()
+        parameters = blueprint.get_parameter_values()
         self.assertEqual(parameters["Param1"], ["us-east-1", "us-west-2"])
 
     def test_resolve_variables_cfn_type_list_invalid_value(self):
@@ -420,32 +453,32 @@ class TestVariables(unittest.TestCase):
             blueprint.resolve_variables(variables)
         variables = blueprint.get_variables()
 
-    def test_get_parameters_cfn_type_list(self):
+    def test_get_parameter_definitions_cfn_type_list(self):
         class TestBlueprint(Blueprint):
             VARIABLES = {
                 "Param1": {"type": EC2AvailabilityZoneNameList},
             }
 
         blueprint = TestBlueprint(name="test", context=MagicMock())
-        parameters = blueprint._get_parameters()
+        parameters = blueprint.get_parameter_definitions()
         self.assertTrue("Param1" in parameters)
         parameter = parameters["Param1"]
         self.assertEqual(parameter["type"],
                          "List<AWS::EC2::AvailabilityZone::Name>")
 
-    def test_get_parameters_cfn_type(self):
+    def test_get_parameter_definitions_cfn_type(self):
         class TestBlueprint(Blueprint):
             VARIABLES = {
                 "Param1": {"type": CFNString},
             }
 
         blueprint = TestBlueprint(name="test", context=MagicMock())
-        parameters = blueprint._get_parameters()
+        parameters = blueprint.get_parameter_definitions()
         self.assertTrue("Param1" in parameters)
         parameter = parameters["Param1"]
         self.assertEqual(parameter["type"], "String")
 
-    def test_required_parameters_cfn_type(self):
+    def test_get_required_parameter_definitions_cfn_type(self):
         class TestBlueprint(Blueprint):
             VARIABLES = {
                 "Param1": {"type": CFNString},
@@ -453,9 +486,10 @@ class TestVariables(unittest.TestCase):
 
         blueprint = TestBlueprint(name="test", context=MagicMock())
         blueprint.setup_parameters()
-        self.assertEqual(blueprint.required_parameters[0][0], "Param1")
+        params = blueprint.get_required_parameter_definitions()
+        self.assertEqual(params.keys()[0], "Param1")
 
-    def test_get_cfn_parameters(self):
+    def test_get_parameter_values(self):
         class TestBlueprint(Blueprint):
             VARIABLES = {
                 "Param1": {"type": int},
@@ -467,6 +501,30 @@ class TestVariables(unittest.TestCase):
         blueprint.resolve_variables(variables)
         variables = blueprint.get_variables()
         self.assertEqual(len(variables), 2)
-        parameters = blueprint.get_cfn_parameters()
+        parameters = blueprint.get_parameter_values()
         self.assertEqual(len(parameters.keys()), 1)
         self.assertEqual(parameters["Param2"], "Value")
+
+    def test_validate_allowed_values(self):
+        allowed_values = ['allowed']
+        valid = validate_allowed_values(allowed_values, "not_allowed")
+        self.assertFalse(valid)
+        valid = validate_allowed_values(allowed_values, "allowed")
+        self.assertTrue(valid)
+
+    def test_blueprint_with_parameters_fails(self):
+        class TestBlueprint(Blueprint):
+            PARAMETERS = {
+                "Param2": {"default": 0, "type": "Integer"},
+            }
+
+        with self.assertRaises(AttributeError):
+            TestBlueprint(name="test", context=MagicMock())
+
+        class TestBlueprint(Blueprint):
+            LOCAL_PARAMETERS = {
+                "Param2": {"default": 0, "type": "Integer"},
+            }
+
+        with self.assertRaises(AttributeError):
+            TestBlueprint(name="test", context=MagicMock())