+++ stacker/tests/test_stacker.py
@@ -8,17 +8,17 @@ class TestStacker(unittest.TestCase):
     def test_stacker_build_parse_args(self):
         stacker = Stacker()
         args = stacker.parse_args(
-            ["build", "-p", "BaseDomain=mike.com", "-r", "us-west-2", "-p",
-             "AZCount=2", "-p", "CidrBlock=10.128.0.0/16",
+            ["build", "-var", "BaseDomain=mike.com", "-r", "us-west-2", "-var",
+             "AZCount=2", "-var", "CidrBlock=10.128.0.0/16",
              "-e", "namespace=test.override",
              "stacker/tests/fixtures/basic.env",
              "stacker/tests/fixtures/vpc-bastion-db-web.yaml"]
         )
-        # verify parameters
-        parameters = args.parameters
-        self.assertEqual(parameters["BaseDomain"], "mike.com")
-        self.assertEqual(parameters["CidrBlock"], "10.128.0.0/16")
-        self.assertEqual(parameters["AZCount"], "2")
+        # verify variables
+        variables = args.variables
+        self.assertEqual(variables["BaseDomain"], "mike.com")
+        self.assertEqual(variables["CidrBlock"], "10.128.0.0/16")
+        self.assertEqual(variables["AZCount"], "2")
         self.assertEqual(args.region, "us-west-2")
         self.assertFalse(args.outline)
         # verify namespace was modified
@@ -27,8 +27,8 @@ class TestStacker(unittest.TestCase):
     def test_stacker_build_context_passed_to_blueprint(self):
         stacker = Stacker()
         args = stacker.parse_args(
-            ["build", "-p", "BaseDomain=mike.com", "-r", "us-west-2", "-p",
-             "AZCount=2", "-p", "CidrBlock=10.128.0.0/16",
+            ["build", "-var", "BaseDomain=mike.com", "-r", "us-west-2", "-var",
+             "AZCount=2", "-var", "CidrBlock=10.128.0.0/16",
              "stacker/tests/fixtures/basic.env",
              "stacker/tests/fixtures/vpc-bastion-db-web.yaml"]
         )
@@ -36,35 +36,33 @@ class TestStacker(unittest.TestCase):
         stacks_dict = args.context.get_stacks_dict()
         blueprint = stacks_dict[args.context.get_fqn("bastion")].blueprint
         self.assertTrue(hasattr(blueprint, "context"))
-        blueprint.create_template()
-        blueprint.setup_parameters()
-        # verify that the bastion blueprint only contains blueprint parameters,
-        # not BaseDomain, AZCount or CidrBlock. Any parameters that get passed
+        blueprint.render_template()
+        # verify that the bastion blueprint only contains blueprint variables,
+        # not BaseDomain, AZCount or CidrBlock. Any variables that get passed
         # in from the command line shouldn't be resovled at the blueprint level
-        self.assertNotIn("BaseDomain", blueprint.parameters)
-        self.assertNotIn("AZCount", blueprint.parameters)
-        self.assertNotIn("CidrBlock", blueprint.parameters)
+        self.assertNotIn("BaseDomain", blueprint.template.parameters)
+        self.assertNotIn("AZCount", blueprint.template.parameters)
+        self.assertNotIn("CidrBlock", blueprint.template.parameters)
 
     def test_stacker_blueprint_property_access_does_not_reset_blueprint(self):
         stacker = Stacker()
         args = stacker.parse_args(
-            ["build", "-p", "BaseDomain=mike.com", "-r", "us-west-2", "-p",
-             "AZCount=2", "-p", "CidrBlock=10.128.0.0/16",
+            ["build", "-var", "BaseDomain=mike.com", "-r", "us-west-2", "-var",
+             "AZCount=2", "-var", "CidrBlock=10.128.0.0/16",
              "stacker/tests/fixtures/basic.env",
              "stacker/tests/fixtures/vpc-bastion-db-web.yaml"]
         )
         stacker.configure(args)
         stacks_dict = args.context.get_stacks_dict()
         bastion_stack = stacks_dict[args.context.get_fqn("bastion")]
-        bastion_stack.blueprint.create_template()
-        bastion_stack.blueprint.setup_parameters()
-        self.assertIn("DefaultSG", bastion_stack.blueprint.parameters)
+        bastion_stack.blueprint.render_template()
+        self.assertIn("DefaultSG", bastion_stack.blueprint.template.parameters)
 
     def test_stacker_build_context_stack_names_specified(self):
         stacker = Stacker()
         args = stacker.parse_args(
-            ["build", "-p", "BaseDomain=mike.com", "-r", "us-west-2", "-p",
-             "AZCount=2", "-p", "CidrBlock=10.128.0.0/16",
+            ["build", "-var", "BaseDomain=mike.com", "-r", "us-west-2", "-var",
+             "AZCount=2", "-var", "CidrBlock=10.128.0.0/16",
              "stacker/tests/fixtures/basic.env",
              "stacker/tests/fixtures/vpc-bastion-db-web.yaml", "--stacks",
              "vpc", "--stacks", "bastion"]
@@ -72,3 +70,15 @@ class TestStacker(unittest.TestCase):
         stacker.configure(args)
         stacks = args.context.get_stacks()
         self.assertEqual(len(stacks), 2)
+
+    def test_stacker_build_fail_when_parameters_in_stack_def(self):
+        stacker = Stacker()
+        args = stacker.parse_args(
+            ["build", "-var", "BaseDomain=mike.com", "-r", "us-west-2", "-var",
+             "AZCount=2", "-var", "CidrBlock=10.128.0.0/16",
+             "stacker/tests/fixtures/basic.env",
+             "stacker/tests/fixtures/vpc-bastion-db-web-pre-1.0.yaml"]
+        )
+        stacker.configure(args)
+        with self.assertRaises(AttributeError):
+            args.context.get_stacks_dict()