+++ stacker/tests/test_util.py
@@ -4,10 +4,14 @@ import string
 import os
 import Queue
 
-from stacker.context import Context
 from stacker.util import (
     cf_safe_name, load_object_from_string,
     camel_to_snake, handle_hooks, retry_with_backoff)
+
+from .factories import (
+    mock_context,
+    mock_provider,
+)
 
 regions = ["us-east-1", "cn-north-1", "ap-northeast-1", "eu-west-1",
            "ap-southeast-1", "ap-southeast-2", "us-west-2", "us-gov-west-1",
@@ -49,7 +53,7 @@ hook_queue = Queue.Queue()
 
 
 def mock_hook(*args, **kwargs):
-    hook_queue.put(args)
+    hook_queue.put(kwargs)
     return True
 
 
@@ -72,39 +76,40 @@ def result_hook(*args, **kwargs):
 class TestHooks(unittest.TestCase):
 
     def setUp(self):
-        self.context = Context({"namespace": "namespace"})
+        self.context = mock_context(namespace="namespace")
+        self.provider = mock_provider(region="us-east-1")
 
     def test_empty_hook_stage(self):
         hooks = []
-        handle_hooks("fake", hooks, "us-east-1", self.context)
+        handle_hooks("fake", hooks, self.provider, self.context)
         self.assertTrue(hook_queue.empty())
 
     def test_missing_required_hook(self):
         hooks = [{"path": "not.a.real.path", "required": True}]
         with self.assertRaises(ImportError):
-            handle_hooks("missing", hooks, "us-east-1", self.context)
+            handle_hooks("missing", hooks, self.provider, self.context)
 
     def test_missing_required_hook_method(self):
         hooks = [{"path": "stacker.hooks.blah", "required": True}]
         with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, "us-east-1", self.context)
+            handle_hooks("missing", hooks, self.provider, self.context)
 
     def test_missing_non_required_hook_method(self):
         hooks = [{"path": "stacker.hooks.blah", "required": False}]
-        handle_hooks("missing", hooks, "us-east-1", self.context)
+        handle_hooks("missing", hooks, self.provider, self.context)
         self.assertTrue(hook_queue.empty())
 
     def test_default_required_hook(self):
         hooks = [{"path": "stacker.hooks.blah"}]
         with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, "us-east-1", self.context)
+            handle_hooks("missing", hooks, self.provider, self.context)
 
     def test_valid_hook(self):
         hooks = [{"path": "stacker.tests.test_util.mock_hook",
                   "required": True}]
-        handle_hooks("missing", hooks, "us-east-1", self.context)
+        handle_hooks("missing", hooks, self.provider, self.context)
         good = hook_queue.get_nowait()
-        self.assertEqual(good[0], "us-east-1")
+        self.assertEqual(good["provider"].region, "us-east-1")
         with self.assertRaises(Queue.Empty):
             hook_queue.get_nowait()
 
@@ -117,15 +122,15 @@ class TestHooks(unittest.TestCase):
         hooks = [{"path": "stacker.tests.test_util.fail_hook",
                   "required": True}]
         with self.assertRaises(SystemExit):
-            handle_hooks("fail", hooks, "us-east-1", self.context)
+            handle_hooks("fail", hooks, self.provider, self.context)
         hooks = [{"path": "stacker.tests.test_util.exception_hook",
                   "required": True}]
         with self.assertRaises(Exception):
-            handle_hooks("fail", hooks, "us-east-1", self.context)
+            handle_hooks("fail", hooks, self.provider, self.context)
         hooks = [{"path": "stacker.tests.test_util.exception_hook",
                   "required": False}]
         # Should pass
-        handle_hooks("ignore_exception", hooks, "us-east-1", self.context)
+        handle_hooks("ignore_exception", hooks, self.provider, self.context)
 
     def test_return_data_hook(self):
         hooks = [